+++ packages/pagination/src/__tests__/Pagination.test.tsx
@@ -107,6 +107,34 @@ describe('Pagination', () => {
       expect(screen.queryByLabelText('Go to page 5')).not.toBeInTheDocument();
       expect(screen.getByLabelText('Go to page 6')).toBeInTheDocument();
     });
+
+    it('should disable buttons in HTML', () => {
+      const { getByLabelText } = render(
+        <Pagination
+          currentPage={1}
+          renderItem={item => <PaginationItem component={'button'} {...item} />}
+          totalPages={1}
+        />,
+      );
+
+      expect(getByLabelText('Go to previous page')).toBeDisabled();
+      expect(getByLabelText('Go to next page')).toBeDisabled();
+    });
+
+    it('should disable non buttons in an accessible way', () => {
+      const { getByLabelText } = render(
+        <Pagination currentPage={1} totalPages={1} />,
+      );
+
+      expect(getByLabelText('Go to previous page')).toHaveAttribute(
+        'aria-disabled',
+        'true',
+      );
+      expect(getByLabelText('Go to next page')).toHaveAttribute(
+        'aria-disabled',
+        'true',
+      );
+    });
   });
 
   it('should have a custom attribute (aria-label in this case) on pagination root element', () => {