+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.1](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/icons@4.2.0...@heycar-uikit/icons@4.2.1) (2023-04-06)
+
+**Note:** Version bump only for package @heycar-uikit/icons
+
+
+
+
+
 # [4.2.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/icons@4.1.0...@heycar-uikit/icons@4.2.0) (2023-03-22)