+++ packages/pagination/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/pagination@1.2.0...@heycar-uikit/pagination@1.2.1) (2023-04-06)
+
+**Note:** Version bump only for package @heycar-uikit/pagination
+
+
+
+
+
+# [1.2.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/pagination@1.1.2...@heycar-uikit/pagination@1.2.0) (2023-03-29)
+
+
+### Features
+
+* **pagination:** fix keys warning and properly disable buttons ([#173](https://github.com/hey-car/heycar-uikit/issues/173)) ([f1c31a4](https://github.com/hey-car/heycar-uikit/commit/f1c31a45f9bedb77829808489db8488a869169c1))
+
+
+
+
+
 ## [1.1.2](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/pagination@1.1.1...@heycar-uikit/pagination@1.1.2) (2023-03-15)
 
 **Note:** Version bump only for package @heycar-uikit/pagination